Change geqdsk constraint for coil fitting
The currents in the coils are not given in G-EQDSK files,
so must be found by fitting. Previously the locations of the
O- and X-points, and some points on the separatrix were used
to constrain this fitting.

This adds a constrain method which uses SciPy.optimize.leastsq
to find the least-squares fit over the entire 2D psi.

This gives better match between mast test cases, and improves
reading of experimental equilibria (faster convergence of nonlinear
iterations)