Apply mask to Jtor to remove currents in SOL
If pprime or ffprime are nonzero at the boundary then
the method used to find toroidal current density Jtor
creates currents in the vacuum

This applies a mask to zero the current in the vacuum,
improving the match to the input current.

When constraining the 2D poloidal flux, the mask is now used
so that only the values in the core are compared.